requirements.txt
python-telegram-bot>=20.0
requests>=2.28.0
Pillow>=10.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
aiofiles>=22.1.0
